requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
argon2-cffi==16.3.0
bleach==2.1.1
cffi==1.11.1              # via argon2-cffi
dj-database-url==0.4.2
django-ckeditor==5.3.0
django-js-asset==0.1.1    # via django-ckeditor
django==1.11.6
faker==0.8.4
flake8==3.4.1
gunicorn==19.7.1
html5lib==1.0b10          # via bleach
mccabe==0.6.1             # via flake8
psycopg2==2.7.3.1
pycodestyle==2.3.1        # via flake8
pycparser==2.18           # via cffi
pyflakes==1.5.0           # via flake8
python-dateutil==2.6.1    # via faker
pytz==2017.2              # via django
raven==6.2.1
shortuuid==0.5.0
six==1.11.0               # via argon2-cffi, bleach, faker, html5lib, python-dateutil
webencodings==0.5.1       # via html5lib
whitenoise==3.3.1
